+++ pkg/firewall/client.go
@@ -6,24 +6,21 @@ import (
 	"strings"
 
 	compute "cloud.google.com/go/compute/apiv1"
-	"github.com/giantswarm/microerror"
 	"github.com/go-logr/logr"
+	"github.com/pkg/errors"
 	computepb "google.golang.org/genproto/googleapis/cloud/compute/v1"
 	"google.golang.org/protobuf/proto"
 	capg "sigs.k8s.io/cluster-api-provider-gcp/api/v1beta1"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
 type Client struct {
 	fwService *compute.FirewallsClient
-	k8sClient client.Client
 	logger    logr.Logger
 }
 
-func NewClient(fwService *compute.FirewallsClient, k8sClient client.Client, logger logr.Logger) *Client {
+func NewClient(fwService *compute.FirewallsClient, logger logr.Logger) *Client {
 	return &Client{
 		fwService: fwService,
-		k8sClient: k8sClient,
 		logger:    logger,
 	}
 }
@@ -66,7 +63,7 @@ func (c *Client) CreateBastionFirewallRule(ctx context.Context, cluster *capg.GC
 			// pass thru, resource already exists
 			return nil
 		} else {
-			return microerror.Mask(err)
+			return errors.WithStack(err)
 		}
 	}
 
@@ -76,7 +73,7 @@ func (c *Client) CreateBastionFirewallRule(ctx context.Context, cluster *capg.GC
 		if isAlreadyExistError(err) {
 			// pass thru, resource already exists
 		} else {
-			return microerror.Mask(err)
+			return errors.WithStack(err)
 		}
 	}
 
@@ -102,14 +99,14 @@ func (c *Client) DeleteBastionFirewallRule(ctx context.Context, cluster *capg.GC
 
 		return nil
 	} else if err != nil {
-		return microerror.Mask(err)
+		return errors.WithStack(err)
 	}
 	err = op.Wait(ctx)
 
 	if isNotFoundError(err) {
 		// pass thru, resource is already deleted
 	} else if err != nil {
-		return microerror.Mask(err)
+		return errors.WithStack(err)
 	}
 
 	logger.Info("Deleted firewall rule for bastion")