+++ CHANGELOG.md
@@ -10,6 +10,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Wait until the `GCPCLuster` has network link in status field.
+### Removed
+
+- No need for a k8s client on the firewall client.
+
+## [0.3.0] - 2022-06-22
+
+### Changed
+
+- Log stack traces when there is an error.
 
 ## [0.2.1] - 2022-06-22
 
@@ -33,7 +42,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Improve logging by adding gcp cluster name being reconciled.
 
-[Unreleased]: https://github.com/giantswarm/capg-firewall-rule-operator/compare/v0.2.1...HEAD
+[Unreleased]: https://github.com/giantswarm/capg-firewall-rule-operator/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/giantswarm/capg-firewall-rule-operator/compare/v0.2.1...v0.3.0
 [0.2.1]: https://github.com/giantswarm/capg-firewall-rule-operator/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/giantswarm/capg-firewall-rule-operator/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/giantswarm/capg-firewall-rule-operator/compare/v0.1.0...v0.1.1